+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## ***CHANGELOG***
 
+### ***1.11.0***
+> - display average daily consumption.
+
+### ***1.1.0***
+> - added a customized Matplotlib NavigationToolBar. 
+
 ### ***1.0.0***
 > - Initital release